Create Wandering in the grid.cpp
/*Write a program that simulates a simple collision detection game. Create a 2D
grid of 3x3 that represents the game world. Each cell can contain either a
reward or a collision object. Randomly populate the array with either – or x,
where – represents a safe place, and x represents a collision object. However,
at the start of the game, the player must be located in the middle of the grid.
So, the character at the middle the grid should be o. If the player moves to a
cell containing a reward, he won, so display the message “Yayy, you won, it
was a reward”. But if that cell contained a collision object, the player will die,
so display the message “You lost, it was a collision object”. Implement
movement of a player within the grid using pointers. The valid moves for
a player are either top, bottom, left, and right.*/